added dynamic filter

+++ src/api/dto/filter.go
@@ -18,12 +18,12 @@ type DynamicFilter struct {
 }
 
 type PagedList[T any] struct {
-	PageNumber      int  `json:"pageNumber"`
-	TotalRows       int  `json:"totalRows"`
-	TotalPages      int  `json:"totalPages"`
-	HasPreviousPage bool `json:"hasPreviousPage"`
-	HasNextPage     bool `json:"hasNextPage"`
-	Items           *[]T `json:"items"`
+	PageNumber      int   `json:"pageNumber"`
+	TotalRows       int64 `json:"totalRows"`
+	TotalPages      int   `json:"totalPages"`
+	HasPreviousPage bool  `json:"hasPreviousPage"`
+	HasNextPage     bool  `json:"hasNextPage"`
+	Items           *[]T  `json:"items"`
 }
 
 type PaginationInput struct {
